third_party/WebKit/LayoutTests/fast/animation/request-animation-frame-missing-arguments.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description('Test how animation methods react to too few arguments.');
shouldThrow("requestAnimationFrame()");
shouldThrow("cancelAnimationFrame()");
</script>
</body>
</html>
